requirements/fastsam.txt
# Base-----------------------------------
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm>=4.64.0

pandas>=1.1.4
seaborn>=0.11.0

gradio==3.35.2

# Ultralytics-----------------------------------
ultralytics == 8.0.120

git+https://github.com/openai/CLIP.git

pycocotools==2.0.6
#https://dl.fbaipublicfiles.com/pytorch3d/packaging/wheels/py38_cu117_pyt201/pytorch3d-0.7.4-cp38-cp38-linux_x86_64.whl
git+https://github.com/facebookresearch/pytorch3d.git
git+https://github.com/cocodataset/panopticapi.git

-e ./video_segmentation/datasets/MVPd
-e ./video_segmentation/models/FastSAM/FastSAM
-e .
